Switch picom to glx backend instead of xrender
Extreme lag was being caused when using the glx backend for picom, since
picom itself has trouble dealing with G-Sync. The proprietary Nvidia
driver ships with a config that blacklists the "compton" process from
using VRR (variable-rate refresh), but when compton changed its name to
picom, this config was never updated. Adding a user config that
explicitly blacklists the "picom" process fixes the issue, and allows
use of the glx backend instead of the less efficient xrender.